server/requirements-dev.txt
Django==5.2.7
djangorestframework==3.16.1
mypy==1.18.2
django-stubs==5.2.8
djangorestframework-stubs==3.16.8
djangorestframework-simplejwt==5.5.1
pillow==12.1.0
dependency-injector==4.48.3
pydantic==2.12.5
pytest==9.0.2
pytest-django==4.11.1